+++ tests/ras/test_ras.py
@@ -7,6 +7,8 @@ from cdislogging import get_logger
 from fence.models import User, UpstreamRefreshToken, GA4GHVisaV1
 from fence.resources.openid.ras_oauth2 import RASOauth2Client as RASClient
 from fence.config import config
+
+from tests.dbgap_sync.conftest import add_visa_manually
 import tests.utils
 
 logger = get_logger(__name__, log_level="debug")
@@ -17,109 +19,6 @@ def add_test_user(db_session, username="admin_user", user_id="5678"):
     db_session.add(test_user)
     db_session.commit()
     return test_user
-
-
-def add_visa_manually(db_session, user, rsa_private_key, kid):
-
-    headers = {"kid": kid}
-
-    decoded_visa = {
-        "iss": "https://stsstg.nih.gov",
-        "sub": "abcde12345aspdij",
-        "iat": int(time.time()),
-        "exp": int(time.time()) + 1000,
-        "scope": "openid ga4gh_passport_v1 email profile",
-        "jti": "jtiajoidasndokmasdl",
-        "txn": "sapidjspa.asipidja",
-        "name": "",
-        "ga4gh_visa_v1": {
-            "type": "https://ras/visa/v1",
-            "asserted": int(time.time()),
-            "value": "https://nig/passport/dbgap",
-            "source": "https://ncbi/gap",
-        },
-    }
-
-    decoded_visa["ras_dbgap_permissions"] = [
-        {
-            "consent_name": "Health/Medical/Biomedical",
-            "phs_id": "phs000991",
-            "version": "v1",
-            "participant_set": "p1",
-            "consent_group": "c1",
-            "role": "designated user",
-            "expiration": "2020-11-14 00:00:00",
-        },
-        {
-            "consent_name": "General Research Use (IRB, PUB)",
-            "phs_id": "phs000961",
-            "version": "v1",
-            "participant_set": "p1",
-            "consent_group": "c1",
-            "role": "designated user",
-            "expiration": "2020-11-14 00:00:00",
-        },
-        {
-            "consent_name": "Disease-Specific (Cardiovascular Disease)",
-            "phs_id": "phs000279",
-            "version": "v2",
-            "participant_set": "p1",
-            "consent_group": "c1",
-            "role": "designated user",
-            "expiration": "2020-11-14 00:00:00",
-        },
-        {
-            "consent_name": "Health/Medical/Biomedical (IRB)",
-            "phs_id": "phs000286",
-            "version": "v6",
-            "participant_set": "p2",
-            "consent_group": "c3",
-            "role": "designated user",
-            "expiration": "2020-11-14 00:00:00",
-        },
-        {
-            "consent_name": "Disease-Specific (Focused Disease Only, IRB, NPU)",
-            "phs_id": "phs000289",
-            "version": "v6",
-            "participant_set": "p2",
-            "consent_group": "c2",
-            "role": "designated user",
-            "expiration": "2020-11-14 00:00:00",
-        },
-        {
-            "consent_name": "Disease-Specific (Autism Spectrum Disorder)",
-            "phs_id": "phs000298",
-            "version": "v4",
-            "participant_set": "p3",
-            "consent_group": "c1",
-            "role": "designated user",
-            "expiration": "2020-11-14 00:00:00",
-        },
-    ]
-
-    encoded_visa = jwt.encode(
-        decoded_visa, key=rsa_private_key, headers=headers, algorithm="RS256"
-    ).decode("utf-8")
-
-    expires = int(decoded_visa["exp"])
-    if user.username == "expired_visa_user":
-        expires -= 100000
-    if user.username == "invalid_visa_user":
-        encoded_visa = encoded_visa[: len(encoded_visa) // 2]
-    if user.username == "TESTUSERD":
-        encoded_visa = encoded_visa[: len(encoded_visa) // 2]
-
-    visa = GA4GHVisaV1(
-        user=user,
-        source=decoded_visa["ga4gh_visa_v1"]["source"],
-        type=decoded_visa["ga4gh_visa_v1"]["type"],
-        asserted=int(decoded_visa["ga4gh_visa_v1"]["asserted"]),
-        expires=expires,
-        ga4gh_visa=encoded_visa,
-    )
-
-    db_session.add(visa)
-    db_session.commit()
 
 
 def add_refresh_token(db_session, user):
